Fix missing return
