+++ packages/protocolbeat/src/store/actions/loadNodes.ts
@@ -1,4 +1,3 @@
-import { merge } from '../../api/merge'
 import type { Node, State } from '../State'
 import { NODE_SPACING, NODE_WIDTH } from '../utils/constants'
 import { recallNodeState } from '../utils/localStore'
@@ -9,44 +8,63 @@ export function loadNodes(
   projectId: string,
   nodes: Node[],
 ): Partial<State> {
-  nodes = merge(state.nodes, nodes)
-  const oldNodes = new Map(state.nodes.map((node) => [node.id, node]))
-  const newIds = new Set(nodes.map((node) => node.id))
-
-  const retainedNodes = state.nodes.filter((node) => newIds.has(node.id))
+  const toAdd: Node[] = nodes.filter(
+    (x) => !state.nodes.some((y) => x.id === y.id),
+  )
+  const existing: Node[] = state.nodes.map((node) => {
+    const newNode = nodes.find((x) => x.id === node.id)
+    return newNode ? { ...newNode, box: node.box, color: node.color } : node
+  })
+  toAdd.push(...createUnknownNodes([...toAdd, ...existing]))
 
   const startX =
-    retainedNodes.length === 0
+    existing.length === 0
       ? 0
-      : Math.max(...retainedNodes.map((node) => node.box.x + node.box.width)) +
+      : Math.max(...existing.map((node) => node.box.x + node.box.width)) +
         NODE_SPACING
 
-  const updatedNodes = nodes
-    .filter((node) => oldNodes.has(node.id))
-    .map((node) => {
-      const oldNode = oldNodes.get(node.id)
-      return oldNode ? { ...node, box: oldNode.box } : node
-    })
-
-  const addedNodes = nodes
-    .filter((node) => !oldNodes.has(node.id))
-    .map((node, i) => {
-      const box = getNodeBoxFromStorage(projectId, node)
-      const x = box?.x ?? startX + (NODE_WIDTH + NODE_SPACING) * i
-      const y = box?.y ?? 0
-      const width = box?.width ?? NODE_WIDTH
-      // height will be updated by updatePositions
-      return { ...node, box: { x, y, width, height: 0 } }
-    })
+  const added = toAdd.map((node, i) => {
+    const box = recallNodeState(projectId)?.locations[node.id]
+    const x = box?.x ?? startX + (NODE_WIDTH + NODE_SPACING) * i
+    const y = box?.y ?? 0
+    const width = box?.width ?? NODE_WIDTH
+    // height will be updated by updatePositions
+    return { ...node, box: { x, y, width, height: 0 } }
+  })
 
   return updateNodePositions({
     ...state,
     projectId,
-    nodes: updatedNodes.concat(addedNodes),
+    nodes: existing.concat(added),
   })
 }
 
-function getNodeBoxFromStorage(projectId: string, node: Node) {
-  const state = recallNodeState(projectId)
-  return state?.locations[node.id]
+function createUnknownNodes(nodes: Node[]): Node[] {
+  const unknownIds = new Set<string>()
+  const knownIds = new Set(nodes.map((node) => node.id))
+
+  for (const node of nodes) {
+    for (const field of node.fields) {
+      if (!knownIds.has(field.target)) {
+        unknownIds.add(field.target)
+      }
+    }
+  }
+
+  return [...unknownIds].map(idToUnknown)
+}
+
+function idToUnknown(id: string): Node {
+  // TODO: better address treatment
+  const address = id.split(':')[1] as string
+  const name = `Unknown ${address.slice(0, 6)}…${address.slice(-4)}`
+  return {
+    id,
+    address,
+    name,
+    box: { x: 0, y: 0, width: 0, height: 0 },
+    color: { l: 0.67, c: 0.166, h: 22 },
+    fields: [],
+    data: null,
+  }
 }
